#6 withFormatSuffix can be configured, false by Default
