+++ src/test/resources/sample.html
@@ -15,15 +15,15 @@
 #### Overview
 Operations about cars
 
-#### 1.1 `/car.{format}/{carId}`
+#### 1.1 `/car/{carId}`
 ##### 1.1.1 getCarById
-**GET** `/car.{format}/{carId}`
+**GET** `/car/{carId}`
 
 Find car by ID
 To get car info by car's Id
 
 ###### URL
-    http://example.com/car.{format}/{carId}
+    http://example.com/car/{carId}
 ###### Parameters
 - path
 
@@ -68,15 +68,15 @@ To get car info by car's Id
 #### Overview
 Operations about garages
 
-#### 2.1 `/garage.{format}/{garageId}`
+#### 2.1 `/garage/{garageId}`
 ##### 2.1.1 getGarageById
-**GET** `/garage.{format}/{garageId}`
+**GET** `/garage/{garageId}`
 
 Find garages by Id
 To get garage info
 
 ###### URL
-    http://example.com/garage.{format}/{garageId}
+    http://example.com/garage/{garageId}
 ###### Parameters
 - path
 
@@ -118,13 +118,13 @@ To get garage info
 
 - - -
 ##### 2.1.2 getGarageById
-**POST** `/garage.{format}/{garageId}`
+**POST** `/garage/{garageId}`
 
 Repair a broken car in garage
 To repair car
 
 ###### URL
-    http://example.com/garage.{format}/{garageId}
+    http://example.com/garage/{garageId}
 ###### Parameters
 - body
 
@@ -185,15 +185,15 @@ To repair car
 #### Overview
 Operations about cars
 
-#### 3.1 `/v2/car.{format}/{carId}`
+#### 3.1 `/v2/car/{carId}`
 ##### 3.1.1 getCarById
-**GET** `/v2/car.{format}/{carId}`
+**GET** `/v2/car/{carId}`
 
 Find car by ID
 To get car info by car's Id
 
 ###### URL
-    http://example.com/v2/car.{format}/{carId}
+    http://example.com/v2/car/{carId}
 ###### Parameters
 - path
